+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added
 
+- Create a `useStepManager` hook to manage step process
 - Create a React `Icon` component that can optionally take alternative text
   for screen reader users.
 - Add a heading for screen reader users on the search results page to have a